Fix critical bug in ReservoirSampler
Basically, we create an array that represents our sample. But the array
is populated with the null value of whatever type underlies the sampler.
What happens is this: if we ask for the nth item, but have seen less
than n items, we will return this nonsense value.

This is unacceptable. This commit will introduce a check at the
beginning of the method that makes sure we've seen enough elements to
return the nth item in the sample