Fix getSampleSet to not return junk
The array that underlies `ReservoirSample` is build to have k elements
initially, and they are populated with some type of null value,
depending on what the underlying type of the sampler is. So, when we
call `getSampleSet`, we normally just return that underlying array,
which means if we've seen < k elements, it will be populated with
garbage null values.

This commit will fix this by returning an array that contains *only* the
values we put into the sampler. That means if we've seen < k elements,
we get back an array containing only the elements that were in the
sampler, and none of the null values.