Add to interface in ReservoirSampler
Add a method that allows us to add an item to the sampler, but that
returns the index into the sample of the item we just added, or -1 if it
did not make it into the sample. This can be used to maintain data that
requires knowledge of what is and is not in the sampler, because if
something gets replaced, it will be given an index that's not -1.