Create method responsible for updating the PF for a new document
Our way of handling this is basically not to handle it. :( This method
will replace the old document update vector with a new one every time we
see a new document. Dramatic improvement, since somehow I wrote this
code to never update a document ever. :(

We also now put document update vectors and topic assignment vectors
into a map that tracks with the reservoir sample updates. This allows us
to resample later with the rejuvenation sequence.