Add method to create ids for every word in the reservoir sample
Basically, a `wordId` is a tuple that contains (docId, wordIdx), where
docId is the index into the reservoir sample that the document occupies,
and wordIdx is the index into that document in which the word lies. By
creating an array of these ids, we are able to simply plug that into the
reservoir which gives us an equiprobable k-sample of those indices.