Fix bug in PfLda.allWordIds that causes us to output too few ids
The bug basically causes us to output, at most, one word id per
vocabulary word. We want to output an id for every word occurrence
though. So if we have a vocab of 3 words, but the number of word
occurrences is 300, then we always output 3 ids. This is highly
undesirable.