Fix really annoying bug in `ParticleStore` instantiation
The `currId` is reset after instantiating the particles. This means that
if a series of particles are given the id 0,1,2,... then currId is set
to 0 after `initParticles` is called, then there will exist at some
point clones with duplicate ids 0,1,2,...

This patch will fix this by moving the statement `currId = 0` to be
above the call to `initParticles`.