Fix nasty assignment bug
We copy the particle during the resampling. Since a particle represents
the state space, this requires copying the topic assignments of the
entire data seen by the particle so far.

This state is stored in two ways. The current document vector is stored
as `particle.currDocVect`. The store of documents from the past is
stored in a hash map backed by a reservoir filter -- that is, a hash map
whose membership is determined by the reservoir filter update rule. This
map is called `particle.rejuvSeqDocVects`.

The bug is as follows.

When we copy the state assignments, we copy the vector of the current
document, but fail to propagate this update back into the hash table.

This commit fixes this awful bug.