Remove pernicious assignment bug.
When we call AssignmentMap.setTopic, one of two things happens. If there
is an entry in the particle for the document `docId`, then we simply set
the word at `wordId` to the given topic.

If, however, there is no entry for `docId`, then we simply replace the
whole particle's entry with a new HashMap containing only an entry for
`docId`. This overwrites anything that was already inside that
particle's entry.

If this sounds stupid and wrong, that's be cause it is both.

This commit will remove this code, and replace it with code that adds a
new entry for `docId` rather than overwriting the entire particle entry.